Port the Kirigami plugin to ECMQmlModule
This ports the main Kirigami plugin to use the new ECMQmlModule.

This removes the need to manually maintain the qmldir file and the qrc
file that is used for static builds. Additionally, the CMake file now
explicitly lists everything that should be installed rather than relying
on things magically being correct.