Introduce "InputMethod" singleton and supporting code to detect virtual keyboard
This adds an InputMethod singleton object that can report the type and
state of the current input method. It also includes some support code in
libkirigami to wrap the KWin VirtualKeyboard interface and uses that to
detect whether a virtual keyboard is currently being used.

While virtual keyboard is the only thing currently supported the intent
is that the InputMethod class can report on more types of input.